Refactor codes in release notes slugs
