+++ components/common/main-content/MainContent.js
@@ -16,28 +16,61 @@
  * under the License.
  */
 
-import * as React from 'react';
-import ReactMarkdown from 'react-markdown';
-import remarkGfm from 'remark-gfm';
-import rehypeRaw from 'rehype-raw';
-import remarkExternalLinks from 'remark-external-links'
+import * as React from "react";
+import ReactMarkdown from "react-markdown";
+import remarkGfm from "remark-gfm";
+import rehypeRaw from "rehype-raw";
+import remarkExternalLinks from "remark-external-links";
+import { getHighlighter } from "shiki";
 
 String.prototype.hashCode = function () {
   var hash = 0,
-    i, chr;
+    i,
+    chr;
   if (this.length === 0) return hash;
   for (i = 0; i < this.length; i++) {
     chr = this.charCodeAt(i);
-    hash = ((hash << 5) - hash) + chr;
+    hash = (hash << 5) - hash + chr;
     hash |= 0;
   }
   return hash;
-}
+};
 
 export default function MainContent(props) {
-
   const content = props.content;
   const codes = props.codes ? new Map(JSON.parse(props.codes)) : new Map();
+
+  // Synatax highlighting
+  const HighlightSyntax = (code, language) => {
+    try {
+      const [codeSnippet, setCodeSnippet] = React.useState([]);
+
+      React.useEffect(() => {
+        async function fetchData() {
+          getHighlighter({
+            theme: "github-light",
+            langs: [
+              "bash",
+              "ballerina",
+              "toml",
+              "yaml",
+              "sh",
+              "json",
+              "graphql",
+              "sql",
+            ],
+          }).then((highlighter) => {
+            setCodeSnippet(highlighter.codeToHtml(code, language));
+          });
+        }
+        fetchData();
+      }, [code, language]);
+
+      return [codeSnippet];
+    } catch {
+      return code;
+    }
+  };
 
   // Add id attributes to headings
   const extractText = (value) => {
@@ -77,13 +110,41 @@ export default function MainContent(props) {
   };
 
   const toc = (show) => {
-    props.handleToc(show)
-  }
+    props.handleToc(show);
+  };
 
   return (
-
     <ReactMarkdown
       components={{
+        h1({ node, inline, className, children, ...props }) {
+          let id = "";
+          toc(true);
+          if (children.length === 1) {
+            id = children[0]
+              .toLowerCase()
+              .replace(/[&\/\\#,+()!$~%.'’":*?<>{}]/g, "")
+              .replace(/ /g, "-");
+          } else {
+            id = scanArray(children);
+          }
+          return (
+            <h1 id={id} data-section={id} className="section">
+              <svg
+                xmlns="http://www.w3.org/2000/svg"
+                width="30"
+                height="30"
+                fill="currentColor"
+                className="bi bi-link-45deg mdButton pe-2"
+                viewBox="0 0 16 16"
+                onClick={(e) => getLink(e.target, id)}
+              >
+                <path d="M4.715 6.542 3.343 7.914a3 3 0 1 0 4.243 4.243l1.828-1.829A3 3 0 0 0 8.586 5.5L8 6.086a1.002 1.002 0 0 0-.154.199 2 2 0 0 1 .861 3.337L6.88 11.45a2 2 0 1 1-2.83-2.83l.793-.792a4.018 4.018 0 0 1-.128-1.287z" />
+                <path d="M6.586 4.672A3 3 0 0 0 7.414 9.5l.775-.776a2 2 0 0 1-.896-3.346L9.12 3.55a2 2 0 1 1 2.83 2.83l-.793.792c.112.42.155.855.128 1.287l1.372-1.372a3 3 0 1 0-4.243-4.243L6.586 4.672z" />
+              </svg>
+              {children}
+            </h1>
+          );
+        },
         h2({ node, inline, className, children, ...props }) {
           let id = "";
           toc(true);
@@ -235,30 +296,49 @@ export default function MainContent(props) {
           );
         },
         code({ node, inline, className, children, ...props }) {
-          if (typeof children[0] === 'string') {
-            const key = (children[0]).trim().split(/\r?\n/).map(row => row.trim()).join('\n');
+          if (typeof children[0] === "string") {
+            const key = children[0]
+              .trim()
+              .split(/\r?\n/)
+              .map((row) => row.trim())
+              .join("\n");
             const highlightedCode = codes.get(key.hashCode());
             if (highlightedCode) {
-              return <div dangerouslySetInnerHTML={{ __html: highlightedCode }} />
+              return (
+                <div dangerouslySetInnerHTML={{ __html: highlightedCode }} />
+              );
             }
           }
 
-          const match = /language-(\w+)/.exec(className || '')
-          return inline ?
+          const match = /language-(\w+)/.exec(className || "");
+          return inline ? (
             <code className={className} {...props}>
               {children}
             </code>
-            : match ?
-              <div dangerouslySetInnerHTML={{ __html: String(children).replace(/\n$/, '') }} />
-              : <pre className='default'>
-                <code className={className} {...props}>
-                  {children}
-                </code>
-              </pre>
+          ) : match ? (
+            <div
+              dangerouslySetInnerHTML={{
+                __html: HighlightSyntax(
+                  String(children).replace(/\n$/, ""),
+                  match[1].toLowerCase()
+                ),
+              }}
+            />
+          ) : (
+            <pre className="default">
+              <code className={className} {...props}>
+                {children}
+              </code>
+            </pre>
+          );
         },
-        table({node, className, children, ...props}) { 
-          return <div className='mdTable'><table {...props}>{children}</table></div>
-        }
+        table({ node, className, children, ...props }) {
+          return (
+            <div className="mdTable">
+              <table {...props}>{children}</table>
+            </div>
+          );
+        },
       }}
       remarkPlugins={[remarkGfm, remarkExternalLinks]}
       rehypePlugins={[rehypeRaw]}
@@ -267,3 +347,6 @@ export default function MainContent(props) {
     </ReactMarkdown>
   );
 }
+
+// content -> data
+// codes -> codeSnippets
