+++ pages/downloads/swan-lake-release-notes/[...slug].js
@@ -19,10 +19,7 @@
 import fs from "fs";
 import matter from "gray-matter";
 import React from "react";
-import ReactMarkdown from "react-markdown";
 import { Container, Col, Button, Offcanvas } from "react-bootstrap";
-import remarkGfm from "remark-gfm";
-import rehypeRaw from "rehype-raw";
 import Image from "next-image-export-optimizer";
 import Head from "next/head";
 
@@ -32,7 +29,7 @@ import { prefix } from "../../../utils/prefix";
 import RNToc from "../../../utils/rl.json";
 import Toc from "../../../components/common/pg-toc/Toc";
 import { highlight } from "../../../utils/highlighter";
-import RenderHeading from "../../../components/common/heading/RenderHeading";
+import MainContent from "../../../components/common/main-content/MainContent";
 
 String.prototype.hashCode = function () {
   var hash = 0,
@@ -109,8 +106,6 @@ export async function getStaticProps({ params: { slug } }) {
 }
 
 export default function PostPage({ frontmatter, content, id, codeSnippets }) {
-  const codes = codeSnippets ? new Map(JSON.parse(codeSnippets)) : new Map();
-
   // Show mobile left nav
   const [show, setShow] = React.useState(false);
   const handleClose = () => setShow(false);
@@ -204,61 +199,11 @@ export default function PostPage({ frontmatter, content, id, codeSnippets }) {
                 </a>
               </Col>
             </div>
-
-            <ReactMarkdown
-              components={{
-                h1: RenderHeading(1, setShowToc),
-                h2: RenderHeading(2, setShowToc),
-                h3: RenderHeading(3, setShowToc),
-                h4: RenderHeading(4, setShowToc),
-                h5: RenderHeading(5, setShowToc),
-                h6: RenderHeading(6, setShowToc),
-                code({ node, inline, className, children, ...props }) {
-                  const key = children[0]
-                    .trim()
-                    .split(/\r?\n/)
-                    .map((row) => row.trim())
-                    .join("\n");
-                  const highlightedCode = codes.get(key.hashCode());
-                  if (highlightedCode) {
-                    return (
-                      <div
-                        dangerouslySetInnerHTML={{ __html: highlightedCode }}
-                      />
-                    );
-                  }
-                  const match = /language-(\w+)/.exec(className || "");
-                  return inline ? (
-                    <code className={className} {...props}>
-                      {children}
-                    </code>
-                  ) : match ? (
-                    <div
-                      dangerouslySetInnerHTML={{
-                        __html: String(children).replace(/\n$/, ""),
-                      }}
-                    />
-                  ) : (
-                    <pre className="default">
-                      <code className={className} {...props}>
-                        {children}
-                      </code>
-                    </pre>
-                  );
-                },
-                table({ node, className, children, ...props }) {
-                  return (
-                    <div className="mdTable">
-                      <table {...props}>{children}</table>
-                    </div>
-                  );
-                },
-              }}
-              remarkPlugins={[remarkGfm]}
-              rehypePlugins={[rehypeRaw]}
-            >
-              {content}
-            </ReactMarkdown>
+            <MainContent
+              content={content}
+              handleToc={setShowToc}
+              codes={codeSnippets}
+            />
           </Container>
         </Col>
         <Col sm={2} xxl={3} className="pageToc d-none d-sm-block">
